Rename with proper capital locale
